requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.5
async-timeout==5.0.1
attrs==25.4.0
black==25.11.0
blinker==1.9.0
click==8.1.8
coverage==7.10.7
exceptiongroup==1.3.1
flake8==7.3.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dotenv==1.2.1
python-Levenshtein==0.27.1
pytokens==0.3.0
RapidFuzz==3.13.0
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
openai==1.58.1
anthropic==0.39.0
ollama==0.4.4
